fix : migracion de los iconos a el footer
